+++ crates/fspy/src/error.rs
@@ -13,7 +13,7 @@ pub enum SpawnError {
         cause: which::Error,
     },
 
-    #[error("failed to initialize seccomp_unotify supervisor: {0}")]
+    #[error("failed to initialize seccomp-filtered ptrace supervisor: {0}")]
     Supervisor(std::io::Error),
 
     #[error("failed to create IPC channel: {0}")]