+++ crates/fspy/src/unix/syscall_handler/getdents.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-use fspy_seccomp_unotify::supervisor::handler::arg::{Caller, Fd};
+use fspy_seccomp_ptrace::supervisor::handler::arg::{Caller, Fd};
 
 use super::SyscallHandler;