+++ crates/fspy_seccomp_ptrace/src/payload/mod.rs
@@ -3,7 +3,7 @@ pub use filter::Filter;
 use wincode::{SchemaRead, SchemaWrite};
 
 #[derive(Debug, SchemaWrite, SchemaRead, Clone)]
-pub struct SeccompPayload {
+pub struct PtracePayload {
     pub(crate) ipc_path: Vec<u8>,
     pub(crate) filter: Filter,
 }