+++ crates/fspy_shared_unix/src/payload.rs
@@ -24,7 +24,7 @@ pub struct Payload {
         not(target_env = "musl"),
         expect(clippy::struct_field_names, reason = "descriptive field name for clarity")
     )]
-    pub seccomp_payload: fspy_seccomp_unotify::payload::SeccompPayload,
+    pub ptrace_payload: fspy_seccomp_ptrace::payload::PtracePayload,
 }
 
 #[cfg(target_os = "macos")]